split queue names into separate file
